--- TournamentManager/TournamentManager/DivisionEnum.cs
@@ -1,20 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace TournamentManager
-{
-    public enum Divisions
-    {
-        Unranked,
-
-        Newbie,
-        Gold,
-        Guardian,
-        Supreme,
-        Diamond,
-        Master
-    }
-}
