+++ TournamentManager/TournamentManager/Exported/Tournaments.txt
@@ -1,12 +1,42 @@
 Katowice 2015
 True
 True
-False
-0
-Atgod|Matus|Valko|0|0|Unranked|-;C0rtana|Andrej|Fukas|0|0|Unranked|-;t1raksvk|Jozef|Kovalak|0|0|Unranked|-;
+True
+4
+Atgod|Matus|Valko|0|0|Unranked|-;C0rtana|Andrej|Fukas|0|0|Unranked|-;t1raksvk|Jozef|Kovalak|0|0|Unranked|-;Saskinka|Saskia|Hrnciarova|0|0|Unranked|-;
 Katowice 2016
 True
 True
 True
-3
-Atgod|Matus|Valko|0|0|Unranked|-;C0rtana|Andrej|Fukas|0|0|Unranked|-;Saskia|Saskia|Hrnciarova|0|0|Unranked|-;
+4
+Atgod|Matus|Valko|0|0|Unranked|-;C0rtana|Andrej|Fukas|0|0|Unranked|-;t1raksvk|Jozef|Kovalak|0|0|Unranked|-;Saskinka|Saskia|Hrnciarova|0|0|Unranked|-;
+Katowice 2017
+True
+True
+True
+4
+Atgod|Matus|Valko|0|0|Unranked|-;C0rtana|Andrej|Fukas|0|0|Unranked|-;t1raksvk|Jozef|Kovalak|0|0|Unranked|-;Saskinka|Saskia|Hrnciarova|0|0|Unranked|-;
+Katowice 2020
+True
+True
+True
+4
+Atgod|Matus|Valko|0|0|Unranked|-;C0rtana|Andrej|Fukas|0|0|Unranked|-;t1raksvk|Jozef|Kovalak|0|0|Unranked|-;Saskinka|Saskia|Hrnciarova|0|0|Unranked|-;
+Katowice 2021
+True
+True
+False
+0
+Atgod|Matus|Valko|0|0|Unranked|-;C0rtana|Andrej|Fukas|0|0|Unranked|-;t1raksvk|Jozef|Kovalak|0|0|Unranked|-;Saskinka|Saskia|Hrnciarova|0|0|Unranked|-;freestajlo|Dominik|Kozubik|0|0|Unranked|-;
+Katovice 2015
+True
+True
+False
+0
+Atgod|Matus|Valko|0|0|Unranked|-;C0rtana|Andrej|Fukas|0|0|Unranked|-;Saskinka|Saskia|Hrnciarova|0|0|Unranked|-;t1raksvk|Jozef|Kovalak|0|0|Unranked|-;freestajlo|Dominik|Kozubik|0|0|Unranked|-;
+Katowice 2000
+True
+True
+False
+0
+Atgod|Matus|Valko|0|0|Unranked|-;C0rtana|Andrej|Fukas|0|0|Unranked|-;freestajlo|Dominik|Kozubik|0|0|Unranked|-;t1raksvk|Jozef|Kovalak|0|0|Unranked|-;Saskinka|Saskia|Hrnciarova|0|0|Unranked|-;
